Tests in justfile and Github Actions update. Lints.

+++ mock-vm/backend/src/mock_vm.rs
@@ -305,7 +305,7 @@ impl MockVmState {
         let events: &Vec<EventData> = events.unwrap();
         let event_position = odra_utils::event_absolute_position(events.len(), index)
             .ok_or(EventError::IndexOutOfBounds)?;
-        Ok(events.get(event_position as usize).unwrap().clone())
+        Ok(events.get(event_position).unwrap().clone())
     }
 
     fn push_callstack_element(&mut self, element: CallstackElement) {
